Update connectivity.txt
Sending a better version for the first command: Using a more direct output format to avoid unnecessary transformations. Give the same result.

In the second command, I am indicating that <root_password> is a parameter. This may be understood by many, but not by those who just want to test the operator. This may be understood by many, but not by those who just want to test the operator.