Update backup-tutorial.md (#100)
Page: "Make a Backup": https://docs.percona.com/percona-operator-for-mysql/pxc/backup-tutorial.html

The documentation does not specify which repository should be used or where it can be found.

In step two, we have instructions to edit the deploy/backup-secret-s3.yaml file. However, there is no prior mention of cloning the repository onto the local machine.